requirements.txt
flask
python-dotenv
soco
yt-dlp